Enhance user profile page with improved layout and document upload functionality
- Updated user information display with a new layout including profile photo, badges, and an information grid.
- Added a modal for uploading documents with validation for file types and size.
- Implemented a new route for handling document uploads in the UserController.

+++ database/migrations/2025_10_15_135000_create_quotation_items_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
             $table->id();
             $table->foreignId('quotation_id')->constrained()->onDelete('cascade');
             $table->foreignId('product_id')->constrained()->onDelete('cascade');
-            $table->integer('quantity');
-            $table->integer('unit_price');
-            $table->integer('subtotal');
+            $table->unsignedBigInteger('quantity');
+            $table->unsignedBigInteger('unit_price');
+            $table->unsignedBigInteger('subtotal');
             $table->timestamps();
         });
     }
